Hypothesis Testing
I decided it would be more appropriate to include the hypothesis testing alongside the visualizations.

List of tests (and others):
- Shapiro-Wilk test for Normality.
- Pearson correlation coefficient.
- Single-variable linear regression.
- F-test to compare population variances.
- t-test to compare population means.

I considered doing a paired test between the same individuals at birth and at day 30, but of course the result of that test for weight and length is obvious. A hypothesis test like that might be useful for your variables, though.